# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce API",
    "description": "REST API for the E-Commerce platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "operationId": "register",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login with email and password",
        "operationId": "login",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns JWT tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get current user profile",
        "operationId": "getProfile",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update user profile (partial update)",
        "operationId": "updateProfile",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/admin": {
      "get": {
        "summary": "Get all users (admin only, paginated)",
        "operationId": "getUsers",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (zero-indexed)",
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "role",
            "in": "query",
            "description": "Filter by user role",
            "schema": { "$ref": "#/components/schemas/UserRoles" }
          },
          {
            "name": "provider",
            "in": "query",
            "description": "Filter by provider",
            "schema": { "$ref": "#/components/schemas/Provider" }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in firstName and email (case-insensitive)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsers"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - ADMIN role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/category": {
      "get": {
        "summary": "Get all categories (paginated, with search)",
        "operationId": "getAllCategories",
        "tags": ["Categories"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (zero-indexed)",
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in categoryName and slug (case-insensitive)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of categories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCategories"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new category (admin only)",
        "operationId": "createCategory",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Category already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - ADMIN role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/category/{slug}": {
      "get": {
        "summary": "Get a category by slug",
        "operationId": "getCategoryBySlug",
        "tags": ["Categories"],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/category/{id}": {
      "put": {
        "summary": "Update a category by ID (admin only)",
        "operationId": "updateCategory",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - ADMIN role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "phoneNumber",
          "dob"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 15,
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 15,
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "example": "secret123"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+123456789"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "example": "1990-01-01"
          },
          "role": {
            "$ref": "#/components/schemas/UserRoles"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "secret123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "description": "All fields are optional — send only the fields you want to update",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 15,
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 15,
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+123456789"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "example": "1990-01-01"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId as hex string",
            "example": "665a1b2c3d4e5f6a7b8c9d0e"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "role": {
            "$ref": "#/components/schemas/UserRoles"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "phoneNumber": {
            "type": "string",
            "description": "Encrypted phone number",
            "example": "+123456789"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "example": "1990-01-01"
          }
        }
      },
      "PaginatedUsers": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 42
              },
              "totalPages": {
                "type": "integer",
                "example": 5
              },
              "page": {
                "type": "integer",
                "example": 0
              },
              "count": {
                "type": "integer",
                "example": 10
              }
            }
          }
        }
      },
      "CategoryRequest": {
        "type": "object",
        "required": ["categoryName"],
        "properties": {
          "categoryName": {
            "type": "string",
            "example": "Electronics"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "665a1b2c3d4e5f6a7b8c9d0e"
          },
          "categoryName": {
            "type": "string",
            "example": "Electronics"
          },
          "slug": {
            "type": "string",
            "example": "electronics"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "description": "User ObjectId (only returned for admins)"
          }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "665a1b2c3d4e5f6a7b8c9d0e"
          },
          "categoryName": {
            "type": "string",
            "example": "Electronics"
          },
          "slug": {
            "type": "string",
            "example": "electronics"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "addedBy": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Admin User"
              }
            }
          }
        }
      },
      "PaginatedCategories": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 10 },
              "totalPages": { "type": "integer", "example": 1 },
              "page": { "type": "integer", "example": 0 },
              "count": { "type": "integer", "example": 10 }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Email already exist"
          },
          "status": {
            "type": "integer",
            "example": 400
          }
        }
      },
      "UserRoles": {
        "type": "string",
        "enum": ["USER", "ADMIN"],
        "default": "USER"
      },
      "Provider": {
        "type": "string",
        "enum": ["SYSTEM", "GOOGLE"],
        "default": "SYSTEM"
      }
    }
  }
}