# src/main/java/com/nazer/e_commerce/category/Dto/CategoryDto.java
package com.nazer.e_commerce.category.Dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CategoryDto {
    @NotBlank(message = "category name is required")
    private String categoryName;
}